LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_click_during_capture_mouse-auxclick-expected.txt


PASS pointerdown/up at child1, no capture
PASS pointerdown/up at child1, capture at child1
FAIL pointerdown/up at child1, capture at child2 assert_equals: expected "pointerdown@child1,gotpointercapture@child2,pointerup@child2,lostpointercapture@child2,auxclick@child2" but got "pointerdown@child1,gotpointercapture@child2,pointerup@child2,lostpointercapture@child2,auxclick@child1"
PASS pointerdown at child1, pointerup at child2, no capture
FAIL pointerdown at child1, pointerup at child2, capture at child1 assert_equals: expected "pointerdown@child1,gotpointercapture@child1,pointerup@child1,lostpointercapture@child1,auxclick@child1" but got "pointerdown@child1,gotpointercapture@child1,pointerup@child1,lostpointercapture@child1,auxclick@parent"
FAIL pointerdown at child1, pointerup at child2, capture at child2 assert_equals: expected "pointerdown@child1,gotpointercapture@child2,pointerup@child2,lostpointercapture@child2,auxclick@child2" but got "pointerdown@child1,gotpointercapture@child2,pointerup@child2,lostpointercapture@child2,auxclick@parent"
